+++ jadx-gui/src/main/java/jadx/gui/settings/JadxSettings.java
@@ -38,8 +38,8 @@ import jadx.cli.LogHelper;
 import jadx.gui.cache.code.CodeCacheMode;
 import jadx.gui.cache.usage.UsageCacheMode;
 import jadx.gui.ui.MainWindow;
+import jadx.gui.ui.action.ActionModel;
 import jadx.gui.ui.codearea.EditorTheme;
-import jadx.gui.ui.menu.ActionModel;
 import jadx.gui.utils.FontUtils;
 import jadx.gui.utils.LafManager;
 import jadx.gui.utils.LangLocale;
