+++ jadx-gui/src/main/java/jadx/gui/settings/ui/shortcut/ShortcutEdit.java
@@ -18,7 +18,7 @@ import javax.swing.UIManager;
 
 import jadx.gui.settings.JadxSettings;
 import jadx.gui.settings.ui.JadxSettingsWindow;
-import jadx.gui.ui.menu.ActionModel;
+import jadx.gui.ui.action.ActionModel;
 import jadx.gui.utils.NLS;
 import jadx.gui.utils.UiUtils;
 import jadx.gui.utils.shortcut.Shortcut;
