+++ jadx-gui/src/main/java/jadx/gui/ui/MainWindow.java
@@ -115,6 +115,8 @@ import jadx.gui.treemodel.JNode;
 import jadx.gui.treemodel.JPackage;
 import jadx.gui.treemodel.JResource;
 import jadx.gui.treemodel.JRoot;
+import jadx.gui.ui.action.ActionModel;
+import jadx.gui.ui.action.JadxGuiAction;
 import jadx.gui.ui.codearea.AbstractCodeArea;
 import jadx.gui.ui.codearea.AbstractCodeContentPanel;
 import jadx.gui.ui.codearea.EditorTheme;
@@ -125,8 +127,6 @@ import jadx.gui.ui.dialog.LogViewerDialog;
 import jadx.gui.ui.dialog.SearchDialog;
 import jadx.gui.ui.filedialog.FileDialogWrapper;
 import jadx.gui.ui.filedialog.FileOpenMode;
-import jadx.gui.ui.menu.ActionModel;
-import jadx.gui.ui.menu.JadxGuiAction;
 import jadx.gui.ui.menu.JadxMenu;
 import jadx.gui.ui.menu.JadxMenuBar;
 import jadx.gui.ui.panel.ContentPanel;
