+++ jadx-gui/src/main/java/jadx/gui/ui/codearea/CommentAction.java
@@ -23,9 +23,9 @@ import jadx.api.metadata.annotations.InsnCodeOffset;
 import jadx.api.metadata.annotations.NodeDeclareRef;
 import jadx.gui.JadxWrapper;
 import jadx.gui.treemodel.JClass;
+import jadx.gui.ui.action.ActionModel;
+import jadx.gui.ui.action.JadxGuiAction;
 import jadx.gui.ui.dialog.CommentDialog;
-import jadx.gui.ui.menu.ActionModel;
-import jadx.gui.ui.menu.JadxGuiAction;
 import jadx.gui.utils.DefaultPopupMenuListener;
 import jadx.gui.utils.NLS;
 import jadx.gui.utils.UiUtils;
