+++ jadx-gui/src/main/java/jadx/gui/ui/codearea/JNodePopupBuilder.java
@@ -3,7 +3,7 @@ package jadx.gui.ui.codearea;
 import javax.swing.JPopupMenu;
 import javax.swing.event.PopupMenuListener;
 
-import jadx.gui.ui.menu.JadxGuiAction;
+import jadx.gui.ui.action.JadxGuiAction;
 import jadx.gui.utils.shortcut.ShortcutsController;
 
 public class JNodePopupBuilder {
